refactor(api): image service -> render service
