--- api/internal/service/image/service.go
@@ -1,7 +1,0 @@
-package image
-
-type ImageService struct{}
-
-func NewImageService() *ImageService {
-	return &ImageService{}
-}
