+++ api/internal/service/render/service.go
@@ -1,0 +1,7 @@
+package render
+
+type RenderService struct{}
+
+func NewRenderService() *RenderService {
+	return &RenderService{}
+}
