+++ api/internal/service/services.go
@@ -4,23 +4,23 @@ import (
 	"dependents.info/internal/config"
 	"dependents.info/internal/service/database"
 	"dependents.info/internal/service/github"
-	"dependents.info/internal/service/image"
+	"dependents.info/internal/service/render"
 )
 
 type Services struct {
 	GitHubOIDCService *github.OIDCService
 	DatabaseService   *database.BadgerService
-	ImageService      *image.ImageService
+	RenderService      *render.RenderService
 }
 
 func BuildAll(cfg *config.Config) *Services {
-	imageService := image.NewImageService()
+	imageService := render.NewRenderService()
 	oidcService := github.NewOIDCService(cfg)
 	dbService := database.NewBadgerService(cfg.DatabasePath)
 
 	return &Services{
 		GitHubOIDCService: oidcService,
 		DatabaseService:   dbService,
-		ImageService:      imageService,
+		RenderService:      imageService,
 	}
 }
